backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2
uvicorn[standard]==0.24.0.post1
gunicorn==21.2.0
openai==1.10.0
slowapi==0.1.9